refactor: 💡 move transformations to separate files
