+++ packages/parser/src/transformations/categorical-chronological/categorical-chronological.test.ts
@@ -332,94 +332,3 @@ test('transform categorical results with "null" values', () => {
     }
   `);
 });
-
-test('transform categorical results in chronological order', () => {
-  const result: IntervalResult[] = [
-    {
-      value: [
-        {
-          author: 'Edwidge Danticat',
-          result: 95,
-          name: 'Love, Anger, Madness',
-        },
-        { author: 'George R. R. Martin', result: 719, name: 'Game of Thrones' },
-      ],
-      timeframe: {
-        start: '2020-01-01T00:00:00.000Z',
-        end: '2020-02-01T00:00:00.000Z',
-      },
-    },
-    {
-      value: [
-        { author: 'Edwidge Danticat', result: 2, name: 'Love, Anger, Madness' },
-        { author: 'George R. R. Martin', result: 11, name: 'Game of Thrones' },
-        { author: 'J.K. Rowling', result: 1, name: 'Harry Potter' },
-      ],
-      timeframe: {
-        start: '2020-02-01T00:00:00.000Z',
-        end: '2020-02-01T16:00:00.000Z',
-      },
-    },
-  ];
-
-  expect(
-    transformCategoricalChronological(
-      {
-        query: {
-          event_collection: 'page_views',
-          analysis_type: 'count',
-          timeframe: {
-            start: '2020-01-01T00:00:00.000Z',
-            end: '2020-02-01T16:00:00.000Z',
-          },
-        },
-        result,
-      },
-      {
-        transformation: 'categorical-chronological',
-        fillEmptyIntervalsKeys: false,
-        mergePropertiesOrder: null,
-      },
-      'table'
-    )
-  ).toMatchInlineSnapshot(`
-    Object {
-      "data": Array [
-        Object {
-          "Interval": "2020-01-01T00:00:00.000Z",
-          "author": "Edwidge Danticat",
-          "count.page_views": 95,
-          "name": "Love, Anger, Madness",
-        },
-        Object {
-          "Interval": "2020-01-01T00:00:00.000Z",
-          "author": "George R. R. Martin",
-          "count.page_views": 719,
-          "name": "Game of Thrones",
-        },
-        Object {
-          "Interval": "2020-02-01T00:00:00.000Z",
-          "author": "Edwidge Danticat",
-          "count.page_views": 2,
-          "name": "Love, Anger, Madness",
-        },
-        Object {
-          "Interval": "2020-02-01T00:00:00.000Z",
-          "author": "George R. R. Martin",
-          "count.page_views": 11,
-          "name": "Game of Thrones",
-        },
-        Object {
-          "Interval": "2020-02-01T00:00:00.000Z",
-          "author": "J.K. Rowling",
-          "count.page_views": 1,
-          "name": "Harry Potter",
-        },
-      ],
-      "keys": Array [
-        "author",
-        "name",
-      ],
-    }
-  `);
-});
