+++ packages/parser/src/transformations/categorical-chronological/categorical-chronological.ts
@@ -1,11 +1,9 @@
 import { Query } from '@keen.io/query';
-import { convertDate } from '@keen.io/time-utils';
 
-import { KEEN_KEY, KEEN_TABLE_INTERVAL } from '../../constants';
+import { defaultTransformation } from './default-transformation';
+import { tableChartTransformation } from './charts';
 
-import { mergePropertiesGroup, fillWithEmptyKeys } from '../../utils';
-
-import { IntervalResult, GroupByResult, ParserSettings } from '../../types';
+import { IntervalResult, ParserSettings } from '../../types';
 
 /**
  * Transforms categorical data with chronological order.
@@ -29,61 +27,19 @@ export const transformCategoricalChronological = (
   }: ParserSettings,
   visualization?: string
 ) => {
-  let data: Record<string, any>[] = [];
-  const keys: Set<string> = new Set();
+  const enableTableFormatter =
+    query?.analysis_type &&
+    query?.event_collection &&
+    visualization === 'table';
 
-  result.forEach((interval) => {
-    const { timeframe, value } = interval as {
-      timeframe: { start: string; end: string };
-      value: GroupByResult[];
-    };
-    const intervalData: Record<string, any> = {};
-    const intervalKeys: Set<string> = new Set();
-
-    const enableTableFormatter =
-      query?.analysis_type &&
-      query?.event_collection &&
-      visualization === 'table';
-
-    if (enableTableFormatter) {
-      const tableValue = `${query?.analysis_type}.${query?.event_collection}`;
-      value.forEach(({ result, ...record }) => {
-        Object.keys(record).forEach((name) => {
-          intervalKeys.add(name as string);
-          intervalData[name] = record[name];
-          intervalData[tableValue] = result;
-        });
-        data.push({
-          [KEEN_TABLE_INTERVAL]: convertDate(timeframe.start, dateModifier),
-          ...intervalData,
-        });
+  return enableTableFormatter
+    ? tableChartTransformation(
+        { query, result },
+        { fillEmptyIntervalsKeys, dateModifier }
+      )
+    : defaultTransformation(result, {
+        mergePropertiesOrder,
+        fillEmptyIntervalsKeys,
+        dateModifier,
       });
-    } else {
-      value
-        .map((groupedProperties) =>
-          mergePropertiesGroup(groupedProperties, mergePropertiesOrder)
-        )
-        .forEach(({ result, ...properties }) => {
-          Object.values(properties).forEach((name) => {
-            intervalKeys.add(name as string);
-            intervalData[name] = result;
-          });
-        });
-
-      data.push({
-        [KEEN_KEY]: convertDate(timeframe.start, dateModifier),
-        ...intervalData,
-      });
-    }
-    intervalKeys.forEach((key) => keys.add(`${key}`));
-  });
-
-  if (fillEmptyIntervalsKeys) {
-    data = fillWithEmptyKeys([...keys], data, 0);
-  }
-
-  return {
-    data,
-    keys: [...keys],
-  };
 };
