+++ packages/parser/src/transformations/chronological/chronological.ts
@@ -1,7 +1,8 @@
 import { Query } from '@keen.io/query';
-import { convertDate } from '@keen.io/time-utils';
 
-import { KEEN_KEY, KEEN_VALUE, KEEN_TABLE_INTERVAL } from '../../constants';
+import { tableChartTransformation } from './charts';
+import { defaultTransformation } from './default-transformation';
+
 import { ParserSettings, IntervalResult } from '../../types';
 
 /**
@@ -22,36 +23,12 @@ export const transformChronological = (
   { dateModifier }: ParserSettings,
   visualization?: string
 ) => {
-  const data: Record<string, any>[] = [];
-  const keys: Set<string> = new Set();
   const enableTableFormatter =
     query?.analysis_type &&
     query?.event_collection &&
     visualization === 'table';
-  const tableValue =
-    enableTableFormatter &&
-    `${query?.analysis_type}.${query?.event_collection}`;
 
-  result.forEach((interval) => {
-    const {
-      value,
-      timeframe: { start: startDate },
-    } = interval as {
-      timeframe: { start: string; end: string };
-      value: number | null | string;
-    };
-    keys.add(enableTableFormatter ? tableValue : KEEN_VALUE);
-    data.push({
-      [enableTableFormatter ? KEEN_TABLE_INTERVAL : KEEN_KEY]: convertDate(
-        startDate,
-        dateModifier
-      ),
-      [enableTableFormatter ? tableValue : KEEN_VALUE]: value,
-    });
-  });
-
-  return {
-    data,
-    keys: [...keys],
-  };
+  return enableTableFormatter
+    ? tableChartTransformation(query, result, dateModifier)
+    : defaultTransformation(result, dateModifier);
 };
