+++ packages/parser/src/transformations/singular/charts/index.ts
@@ -1,0 +1,3 @@
+import { tableChartTransformation } from './table-transformation';
+
+export { tableChartTransformation };
