+++ packages/parser/src/transformations/singular/singular.ts
@@ -1,14 +1,9 @@
 import { Query } from '@keen.io/query';
 
+import { tableChartTransformation } from './charts';
 import { ParserSettings } from '../../types';
 
-import {
-  DEFAULT_NAME,
-  KEEN_KEY,
-  KEEN_VALUE,
-  KEEN_TABLE_VALUE,
-  KEEN_TABLE_ANALYSIS,
-} from '../../constants';
+import { DEFAULT_NAME, KEEN_KEY, KEEN_VALUE } from '../../constants';
 
 /**
  * Transforms singular data.
@@ -35,16 +30,7 @@ export const transformSingular = (
     query?.event_collection &&
     visualization === 'table'
   ) {
-    const { analysis_type, event_collection } = query;
-    return {
-      data: [
-        {
-          [KEEN_TABLE_ANALYSIS]: `${analysis_type}.${event_collection}`,
-          [KEEN_TABLE_VALUE]: value,
-        },
-      ],
-      keys: [KEEN_TABLE_VALUE],
-    };
+    return tableChartTransformation(query, value);
   }
 
   return {
